Lab 6 - Cannot get categories and elements

+++ knife-store/src/main/java/pl/edu/pg/eti/kask/store/knife/service/KnifeService.java
@@ -1,46 +1,82 @@
 package pl.edu.pg.eti.kask.store.knife.service;
 
-import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.ejb.EJBAccessException;
+import jakarta.ejb.LocalBean;
+import jakarta.ejb.Stateless;
 import jakarta.inject.Inject;
-import jakarta.transaction.Transactional;
+import jakarta.security.enterprise.SecurityContext;
 import lombok.NoArgsConstructor;
 import pl.edu.pg.eti.kask.store.knife.entity.Knife;
 import pl.edu.pg.eti.kask.store.knife.repository.api.CategoryRepository;
 import pl.edu.pg.eti.kask.store.knife.repository.api.KnifeRepository;
 import pl.edu.pg.eti.kask.store.user.entity.User;
+import pl.edu.pg.eti.kask.store.user.entity.UserRoles;
 import pl.edu.pg.eti.kask.store.user.repository.api.UserRepository;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-@ApplicationScoped
+@LocalBean
+@Stateless
 @NoArgsConstructor(force = true)
 public class KnifeService {
     private final KnifeRepository knifeRepository;
     private final CategoryRepository categoryRepository;
     private final UserRepository userRepository;
+    private final SecurityContext securityContext;
 
     @Inject
-    public KnifeService(KnifeRepository knifeRepository, CategoryRepository categoryRepository, UserRepository userRepository, CategoryService categoryService) {
+    public KnifeService(KnifeRepository knifeRepository, CategoryRepository categoryRepository, UserRepository userRepository, CategoryService categoryService,  @SuppressWarnings("CdiInjectionPointsInspection") SecurityContext securityContext) {
         this.knifeRepository = knifeRepository;
         this.categoryRepository = categoryRepository;
         this.userRepository = userRepository;
+        this.securityContext = securityContext;
     }
 
+    @RolesAllowed(UserRoles.USER)
     public Optional<Knife> find(UUID id){
         return knifeRepository.find(id);
     }
 
+    @RolesAllowed(UserRoles.USER)
     public Optional<Knife> find(User user, UUID id){
         return knifeRepository.findByIdAndUser(id, user);
     }
 
+    @RolesAllowed(UserRoles.USER)
     public List<Knife> findAll(){
         return knifeRepository.findAll();
     }
 
-    @Transactional
+    @RolesAllowed(UserRoles.USER)
+    public List<Knife> findAll(User user) {
+        return knifeRepository.findAllByUser(user);
+    }
+
+    @RolesAllowed(UserRoles.USER)
+    public Optional<Knife> findForCallerPrincipal(UUID id) {
+        if (securityContext.isCallerInRole(UserRoles.ADMIN)) {
+            return find(id);
+        }
+        User user = userRepository.findByLogin(securityContext.getCallerPrincipal().getName())
+                .orElseThrow(IllegalStateException::new);
+        return find(user, id);
+    }
+
+    @RolesAllowed(UserRoles.USER)
+    public List<Knife> findAllForCallerPrincipal() {
+        if (securityContext.isCallerInRole(UserRoles.ADMIN)) {
+            return findAll();
+        }
+        User user = userRepository.findByLogin(securityContext.getCallerPrincipal().getName())
+                .orElseThrow(IllegalStateException::new);
+        return findAll(user);
+    }
+
+
+    @RolesAllowed(UserRoles.USER)
     public void create(Knife knife){
 //        if(knifeRepository.find(knife.getId()).isPresent()){
 //            throw new IllegalArgumentException("Knife with given id already exists");
@@ -51,26 +87,53 @@ public class KnifeService {
         knifeRepository.create(knife);
     }
 
-    @Transactional
+    @RolesAllowed(UserRoles.USER)
+    public void createForCallerPrincipal(Knife knife) {
+        User user = userRepository.findByLogin(securityContext.getCallerPrincipal().getName())
+                .orElseThrow(IllegalStateException::new);
+
+        knife.setUser(user);
+        knifeRepository.create(knife);
+    }
+
+    @RolesAllowed(UserRoles.USER)
     public void update(Knife knife){
+        checkAdminRoleOrOwner(knifeRepository.find(knife.getId()));
         knifeRepository.update(knife);
     }
 
-    @Transactional
+    @RolesAllowed(UserRoles.USER)
     public void delete(Knife knife){
+        checkAdminRoleOrOwner(knifeRepository.find(knife.getId()));
         knifeRepository.delete(knife);
     }
 
-    @Transactional
+    @RolesAllowed(UserRoles.USER)
     public void delete(UUID id){
+        checkAdminRoleOrOwner(knifeRepository.find(id));
         knifeRepository.delete(knifeRepository.find(id).orElseThrow());
     }
 
+    @RolesAllowed(UserRoles.USER)
     public Optional<List<Knife>> findAllByCategory(UUID id) {
         return categoryRepository.find(id).map(knifeRepository::findAllByCategory);
     }
 
+    @RolesAllowed(UserRoles.USER)
     public Optional<List<Knife>> findAllByUser(UUID id) {
         return userRepository.find(id).map(knifeRepository::findAllByUser);
     }
+
+    private void checkAdminRoleOrOwner(Optional<Knife> reservation) throws EJBAccessException {
+        if (securityContext.isCallerInRole(UserRoles.ADMIN)) {
+            return;
+        }
+        if (securityContext.isCallerInRole(UserRoles.USER)
+                && reservation.isPresent()
+                && reservation.get().getUser().getLogin().equals(securityContext.getCallerPrincipal().getName())) {
+            return;
+        }
+        throw new EJBAccessException("Caller not authorized.");
+    }
+
 }
